Complete re-write of the metadata extraction process. Now done with key and value arrays. This allows for easier error checking and allows for modification in the future.
The process now will check if it was able to extract the metadata from the input audio file.
If it can't, it will ask the user to input the data manually.

Changed the way that the COMMENT metadata tag is created. It is generated from additional values in the array.

Also added in the ability to initiate a manual override of the metadata extraction.
Running with flag -m will allow the user to enter all the details themself.